requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
curl_cffi==0.11.1
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.1
MarkupSafe==3.0.2
multidict==6.4.4
pillow==11.2.1
propcache==0.3.1
PyCharacterAI==2.2.46
pycparser==2.22
PyNaCl==1.5.0
pyScss==1.4.0
python-dotenv==1.1.0
RapidFuzz==3.13.0
requests==2.32.3
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0